Make some changes to the model structure
Move payee to be on the transaction.
Add memo to both transaction and entry.
Add date to the transaction.

Don't make it a datetime since the date will often be entered in
manually, and further refinement of the transaction moment is probably
unnecessary.

Also don't add it to the entry, but only the transaction. It logically
occurs at the same time, so even though different places may have a
different date in their records, it seems reasonable to force that a
specific date is agreed upon for the whole transaction.