Update models to make more fexible
In accounting, there are journals and sub-journals, accounts and
sub-accounts. In budgeting, there are categories and sub-categories.

Give some kind of ID field to all addressable models. Some, because of
their relationships, may not be able to have a unique constraint on them
at the database / ORM layer, and will have to be enforced by custom save
methods and/or other devices.